:sparkles: feat: change main page
