+++ components/layout/Header.tsx
@@ -16,7 +16,7 @@ const Header: React.FC = () => {
           </Link>
         </div>
         <div className="button-container">
-          <Link href="/about">지역&카테고리 검색</Link>
+          <Link href="/result">지역&카테고리 검색</Link>
           <Link href="/map">지도 검색</Link>
           <Link href="/about">케이크크 스토리</Link>
           {/* <button className="toggle-button my-page">
