--- components/toy/index.ts
@@ -1,3 +1,0 @@
-import ToyGrid from './ToyGrid'
-import ToyBanner from './ToyBanner'
-export { ToyGrid, ToyBanner }
