+++ pages/index.tsx
@@ -1,46 +1,117 @@
-import type { GetServerSideProps, GetStaticProps } from 'next'
-import Head from 'next/head'
-import BannerContainer from '../components/main/banner'
+import type { GetStaticProps } from 'next'
 import Layout from '../components/layout'
-import SeacrhContainer from '../components/main/search/SeacrhContainer'
-import RankContainer from '../components/main/rank-section/RankContainer'
-import AreaContainer from '../components/main/area-section'
-import * as Styled from '../styles/home'
-import { CategoryProvider, LocationProvider } from '../context'
 
-import { fetchPopular, ItemResponseProps } from '../src/api/api'
+import { ItemResponseProps } from '../src/api/api'
+import { PlayBanner } from '../components/play'
+import { motion } from 'framer-motion'
+import MainPlayItem from '../components/play/item/MainPlayItem'
+import ImagePlayItem from '../components/play/item/ImagePlayItem'
+import DetailPlayItem from '../components/play/item/DetailPlayItem'
+import AnimatePlayItem from '../components/play/item/AnimatePlayItem'
+
+const columnVariants1 = {
+  hidden: {
+    width: '0vw',
+    y: '100%',
+  },
+  visible: {
+    width: '30vw',
+    y: '0%',
+    transition: {
+      delay: 3,
+      width: {
+        duration: 0,
+      },
+      y: {
+        duration: 1,
+      },
+    },
+  },
+}
+
+const columnVariants2 = {
+  hidden: {
+    width: '0vw',
+    y: '-100%',
+  },
+  visible: {
+    width: '30vw',
+    y: '0%',
+    transition: {
+      delay: 3,
+      width: {
+        duration: 0,
+      },
+      y: {
+        duration: 1,
+      },
+    },
+  },
+}
 
 const Home = ({ data }: { data: ItemResponseProps[] }) => {
   return (
     <Layout>
-      <Head>
-        <title>🎂 케이크크</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <Styled.Home>
-        <LocationProvider>
-          <CategoryProvider>
-            <BannerContainer />
-            <SeacrhContainer />
-            <RankContainer data={data} />
-            {/* <AreaContainer /> */}
-          </CategoryProvider>
-        </LocationProvider>
-      </Styled.Home>
+      <div style={{ display: 'flex' }}>
+        <PlayBanner />
+        <motion.div
+          variants={columnVariants1}
+          initial="hidden"
+          animate="visible"
+        >
+          <MainPlayItem />
+          <ImagePlayItem
+            mainImage="/img/Rectangle 617.png"
+            heading="No 밀가루 Only 국산 쌀가루<br/>베이커리 숲에서 건강한 케이크를 맛보세요"
+          />
+          <ImagePlayItem
+            mainImage="/img/Rectangle 621.png"
+            heading="귀여운 입체 케이크를 찾는다면?<br/>
+          수이아로 오세요!"
+          />
+        </motion.div>
+        <motion.div
+          variants={columnVariants2}
+          initial="hidden"
+          animate="visible"
+        >
+          <DetailPlayItem
+            mainImage="/img/Rectangle 614.png"
+            heading={'힙지로 감성,<br/>이제 금붕어에서 느껴보세요'}
+            data={[
+              {
+                id: 1,
+                image: '/img/Rectangle 615.png',
+                submain: '홀케이크',
+                main: '빈티지 거울 케이크',
+                tags: '#중구 #레터링 케이크',
+              },
+              {
+                id: 2,
+                image: '/img/Rectangle 616.png',
+                submain: '하트케이크',
+                main: '하트 포토케이크',
+                tags: '#중구 #포토 케이크',
+              },
+            ]}
+          />
+          <AnimatePlayItem />
+        </motion.div>
+      </div>
     </Layout>
   )
 }
 
-export const getStaticProps: GetStaticProps = async () => {
-  try {
-    const data = await fetchPopular()
-    return {
-      props: { data }, // will be passed to the page component as props
-    }
-  } catch (err) {
-    console.error('Failed to fetch data in /', err)
-    return { props: { data: [] } }
+export const getStaticProps: GetStaticProps = async (context) => {
+  // const data = await fetchPopular()
+  // const response = await fetch('http://localhost:3000/api/cakestore/popular')
+  // const data = await response.json()
+  // console.log('data fetch', data)
+
+  const data = [] as ItemResponseProps[]
+
+  return {
+    props: { data }, // will be passed to the page component as props
   }
 }
